+++ webpack/static-pages/webpack.prod.config.js
@@ -2,6 +2,7 @@ const HtmlWebpackPlugin = require('html-webpack-plugin');
 const path = require('path');
 const webpackFiles = require('../../lib/build/files/webpack_files');
 const Package = require('./../../package.json');
+const webpack = require('webpack');
 
 const VERSION = Package.version;
 
@@ -13,13 +14,19 @@ module.exports = {
     publicPath: '/assets/'
   },
   devtool: 'source-map',
-  plugins: Object.keys(webpackFiles).map((entryName) => {
-    return new HtmlWebpackPlugin({
-      inject: false,
-      cache: false,
-      filename: path.resolve(__dirname, `../../public/static/${entryName}/index.html`),
-      template: path.resolve(__dirname, '../../lib/assets/javascripts/dashboard/statics/index.jst.ejs'),
-      config: webpackFiles[entryName]
-    });
-  })
+  plugins: [
+    new webpack.DefinePlugin({
+      VERSION: `'${VERSION}'`
+    }),
+
+    ...Object.keys(webpackFiles).map((entryName) => {
+      return new HtmlWebpackPlugin({
+        inject: false,
+        cache: false,
+        filename: path.resolve(__dirname, `../../public/static/${entryName}/index.html`),
+        template: path.resolve(__dirname, '../../lib/assets/javascripts/dashboard/statics/index.jst.ejs'),
+        config: webpackFiles[entryName]
+      });
+    })
+  ]
 };